Move Migration into namespace and remove old
This may be a bit controversial, as it's technically backwards breaking,
but I don't think this matters, given how long the plugin has been on
.org, I don't expect this to actually affect anyone.
