+++ app/Providers/CoreServiceProvider.php
@@ -26,13 +26,6 @@ class CoreServiceProvider implements ServiceProvider {
 			->define( 'register.data', new Data )
 			->define( 'facade.adapter', new Adapter )
 			->define( 'facade.database', new Database( $container->fetch( 'facade.adapter' ) ) )
-			->define( 'migration', new Migration(
-				$container->fetch( 'facade.database' ),
-				$container->fetch( 'facade.adapter' ),
-				$container->fetch( 'database' ),
-				$container->fetch( 'slug' ),
-				$container->fetch( 'version' )
-			) )
 			->define(
 				'save',
 				new Save(
