+++ app/Providers/ListenerServiceProvider.php
@@ -2,6 +2,7 @@
 namespace Intraxia\Gistpen\Providers;
 
 use Intraxia\Gistpen\Listener\Database;
+use Intraxia\Gistpen\Listener\Migration;
 use Intraxia\Gistpen\Listener\Sync;
 use Intraxia\Jaxion\Contract\Core\Container;
 use Intraxia\Jaxion\Contract\Core\ServiceProvider;
@@ -23,6 +24,15 @@ class ListenerServiceProvider implements ServiceProvider {
 
 		$container->define( 'listener.sync', function( Container $container ) {
 			return new Sync( $container->fetch( 'jobs' ) );
+		} );
+
+		$container->define( 'listener.migration', function( Container $container ) {
+			return new Migration(
+				$container->fetch( 'database' ),
+				'wpgp',
+				$container->fetch( 'slug' ),
+				$container->fetch( 'version' )
+			);
 		} );
 	}
 }
